@@ -9,35 +9,33 @@
     ],
     "license": "GPL-3.0-or-later",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3",
         "ext-gd": "*",
         "ext-json": "*",
         "ext-zip": "*",
         "chillerlan/php-qrcode": "^3.2",
         "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^1.0",
+        "fruitcake/laravel-cors": "^2.0",
         "funkjedi/composer-include-files": "^1.0",
         "google/recaptcha": "^1.2",
-        "guzzlehttp/guzzle": "^6.3",
-        "laravel/framework": "^7.0",
-        "laravel/socialite": "^4.4",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.0",
+        "laravel/socialite": "^5.0",
         "laravel/tinker": "^2.0",
         "laravel/ui": "^2.0",
         "pragmarx/google2fa": "^8.0",
         "symfony/filesystem": "^5.0",
         "textalk/websocket": "^1.4",
         "thedudeguy/rcon": "^1.0",
-        "truckersmp/steam-socialite": "^1.0",
         "xpaw/php-source-query-class": "^1.0"
     },
     "require-dev": {
         "barryvdh/laravel-debugbar": "^3.2",
-        "beyondcode/laravel-dump-server": "^1.4",
-        "facade/ignition": "^2.0",
+        "facade/ignition": "^2.3.6",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3"
     },
     "suggest": {
         "wikimedia/composer-merge-plugin": "This allow to merge the composer.json of the plugins in the main composer autoload"
@@ -67,14 +65,12 @@
     },
     "autoload": {
         "psr-4": {
-            "Azuriom\\": "app/"
+            "Azuriom\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         },
         "files": [
             "app/helpers.php"
-        ],
-        "classmap": [
-            "database/seeds",
-            "database/factories"
         ]
     },
     "autoload-dev": {
