@@ -3,6 +3,7 @@
 namespace Azuriom\Exceptions;
 
 use Azuriom\Azuriom;
+use Exception;
 use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\ViewErrorBag;
@@ -31,32 +32,15 @@ class Handler extends ExceptionHandler
     ];
 
     /**
-     * Report or log an exception.
+     * Register the exception handling callbacks for the application.
      *
-     * @param  \Throwable  $exception
      * @return void
-     *
-     * @throws \Exception
      */
-    public function report(Throwable $exception)
+    public function register()
     {
-        parent::report($exception);
-
-        $this->reportException($exception);
-    }
-
-    /**
-     * Render an exception into an HTTP response.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \Throwable  $exception
-     * @return \Symfony\Component\HttpFoundation\Response
-     *
-     * @throws \Throwable
-     */
-    public function render($request, Throwable $exception)
-    {
-        return parent::render($request, $exception);
+        $this->reportable(function (Exception $e) {
+            //
+        });
     }
 
     /**
@@ -112,15 +96,12 @@ class Handler extends ExceptionHandler
 
     /**
      * Report the exception to Azuriom to provide quick fix of errors.
+     *
      * @param  \Throwable  $exception
      */
     protected function reportException(Throwable $exception)
     {
-        if ($this->shouldntReport($exception)) {
-            return;
-        }
-
-        if (config('app.debug')) {
+        if (config('app.debug') || app()->runningInConsole()) {
             return;
         }
 
