@@ -2,22 +2,18 @@
 
 namespace Azuriom\Providers;
 
+use Illuminate\Cache\RateLimiting\Limit;
 use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\RateLimiter;
 use Illuminate\Support\Facades\Route;
 
 class RouteServiceProvider extends ServiceProvider
 {
     /**
-     * This namespace is applied to your controller routes.
-     *
-     * In addition, it is set as the URL generator's root namespace.
-     *
-     * @var string
-     */
-    protected $namespace = 'Azuriom\Http\Controllers';
-
-    /**
      * The path to the "home" route for your application.
+     *
+     * This is used by Laravel authentication to redirect users after login.
      *
      * @var string
      */
@@ -34,7 +30,21 @@ class RouteServiceProvider extends ServiceProvider
         Route::pattern('slug', '[a-z0-9-]+');
         Route::pattern('name', '[a-z0-9_-]+');
 
-        parent::boot();
+        $this->configureRateLimiting();
+
+        $this->routes(function () {
+            Route::prefix('admin')
+                ->middleware(['web', 'auth', 'can:admin.access', 'verified'])
+                ->name('admin.')
+                ->group(base_path('routes/admin.php'));
+
+            Route::prefix('api')
+                ->middleware('api')
+                ->group(base_path('routes/api.php'));
+
+            Route::middleware('web')
+                ->group(base_path('routes/web.php'));
+        });
     }
 
     /**
@@ -52,45 +62,14 @@ class RouteServiceProvider extends ServiceProvider
     }
 
     /**
-     * Define the "web" routes for the application.
-     *
-     * These routes all receive session state, CSRF protection, etc.
+     * Configure the rate limiters for the application.
      *
      * @return void
      */
-    protected function mapWebRoutes()
+    protected function configureRateLimiting()
     {
-        Route::middleware('web')
-            ->namespace($this->namespace)
-            ->group(base_path('routes/web.php'));
-    }
-
-    /**
-     * Define the "admin" routes for the application.
-     *
-     * @return void
-     */
-    protected function mapAdminRoutes()
-    {
-        Route::prefix('admin')
-            ->middleware(['web', 'auth', 'can:admin.access', 'verified'])
-            ->namespace($this->namespace.'\Admin')
-            ->name('admin.')
-            ->group(base_path('routes/admin.php'));
-    }
-
-    /**
-     * Define the "api" routes for the application.
-     *
-     * These routes are typically stateless.
-     *
-     * @return void
-     */
-    protected function mapApiRoutes()
-    {
-        Route::prefix('api')
-            ->middleware('api')
-            ->namespace($this->namespace.'\Api')
-            ->group(base_path('routes/api.php'));
+        RateLimiter::for('api', function (Request $request) {
+            return Limit::perMinute(60);
+        });
     }
 }
