@@ -90,9 +90,11 @@ return [
     | Session Cache Store
     |--------------------------------------------------------------------------
     |
-    | When using the "apc", "memcached", or "dynamodb" session drivers you may
+    | While using one of the framework's cache driven session backends you may
     | list a cache store that should be used for these sessions. This value
     | must match with one of the application's configured cache "stores".
+    |
+    | Affects: "apc", "dynamodb", "memcached", "redis"
     |
     */
 
@@ -122,7 +124,7 @@ return [
     |
     */
 
-    'cookie' => 'azuriom_session',
+    'cookie' => env('SESSION_COOKIE', 'azuriom_session'),
 
     /*
     |--------------------------------------------------------------------------
