@@ -1,10 +1,12 @@
 <?php
 
+namespace Database\Seeders;
+
 use Azuriom\Models\Permission;
 use Azuriom\Models\Role;
 use Illuminate\Database\Seeder;
 
-class PermissionsTableSeeder extends Seeder
+class PermissionSeeder extends Seeder
 {
     /**
      * Run the database seeds.
